# ✨ Improve overall included keywords performance