--- a/sds/src/scanner/mod.rs
+++ b/sds/src/scanner/mod.rs
@@ -76,6 +76,8 @@ pub trait CompiledRuleDyn: Send + Sync {
         match_emitter: &mut dyn MatchEmitter,
         true_positive_rule_idx: &[usize],
         scanner_labels: &Labels,
+        path: &Path,
+        should_kws_match_event_paths: bool,
     );
 
     // Whether a match from this rule should be excluded (marked as a false-positive)
@@ -120,6 +122,8 @@ impl<T: CompiledRule> CompiledRuleDyn for T {
         match_emitter: &mut dyn MatchEmitter,
         true_positive_rule_idx: &[usize],
         scanner_labels: &Labels,
+        path: &Path,
+        should_kws_match_event_paths: bool,
     ) {
         let group_data_any = group_data
             .entry(TypeId::of::<T::GroupData>())
@@ -133,6 +137,8 @@ impl<T: CompiledRule> CompiledRuleDyn for T {
             excluded_matches,
             match_emitter,
             true_positive_rule_idx,
+            path,
+            should_kws_match_event_paths,
         )
     }
 
@@ -178,6 +184,8 @@ pub trait CompiledRule: Send + Sync {
         excluded_matches: &mut AHashSet<String>,
         match_emitter: &mut dyn MatchEmitter,
         true_positive_rule_idx: &[usize],
+        path: &Path,
+        should_kws_match_event_paths: bool,
     );
 
     // Whether a match from this rule should be excluded (marked as a false-positive)
@@ -699,6 +707,10 @@ impl<'a, E: Encoding> ContentVisitor<'a> for ScannerContentVisitor<'a, E> {
                     &mut emitter,
                     true_positive_rule_idx,
                     &self.scanner.labels,
+                    path,
+                    self.scanner
+                        .scanner_features
+                        .should_keywords_match_event_paths,
                 );
             }
         });
@@ -845,6 +857,8 @@ mod test {
             _excluded_matches: &mut AHashSet<String>,
             match_emitter: &mut dyn MatchEmitter,
             _true_positive_rule_idx: &[usize],
+            _path: &Path,
+            _should_kws_match_event_paths: bool,
         ) {
             match_emitter.emit(StringMatch { start: 10, end: 16 });
         }