--- a/sds/src/scanner/regex_rule/compiled.rs
+++ b/sds/src/scanner/regex_rule/compiled.rs
@@ -205,7 +205,7 @@ impl RegexCompiledRule {
                 }
             };
 
-            if has_verified_kws_in_path.is_none_or(|x| x == false) {
+            if has_verified_kws_in_path.is_none_or(|x| !x) {
                 // We don't deal with true positives is in this case, because keywords don't match the path.
                 // Return early.
                 return;