--- a/sds/src/scanner/regex_rule/compiled.rs
+++ b/sds/src/scanner/regex_rule/compiled.rs
@@ -123,24 +123,6 @@ impl RegexCompiledRule {
         path: &Path,
         should_kws_match_event_paths: bool,
     ) {
-        if should_kws_match_event_paths {
-            let sanitized_path = path.sanitize();
-            if contains_keyword_in_path(&sanitized_path, &included_keywords.keywords_pattern) {
-                // since the path contains a match, we can skip future included keyword checks
-                let true_positive_search = self.true_positive_matches(
-                    content,
-                    0,
-                    regex_caches.get(&self.regex),
-                    false,
-                    exclusion_check,
-                    excluded_matches,
-                );
-                for string_match in true_positive_search {
-                    match_emitter.emit(string_match);
-                }
-                return;
-            }
-        }
         if true_positive_rule_idx.contains(&self.rule_index) {
             // since the path contains a match, we can skip future included keyword checks
             let true_positive_search = self.true_positive_matches(
@@ -158,6 +140,35 @@ impl RegexCompiledRule {
         }
 
         let mut included_keyword_matches = included_keywords.keyword_matches(content);
+
+        if included_keyword_matches.is_empty(regex_caches) {
+            if should_kws_match_event_paths {
+                let true_positive_search = self.true_positive_matches(
+                    content,
+                    0,
+                    regex_caches.get(&self.regex),
+                    false,
+                    exclusion_check,
+                    excluded_matches,
+                );
+
+                let mut has_kws_in_path: Option<bool> = None;
+                for string_match in true_positive_search {
+                    // If has_kws_in_path is None, calculate it. We only do it once.
+                    if has_kws_in_path.is_none() {
+                        has_kws_in_path = Some(contains_keyword_in_path(
+                            &path.sanitize(),
+                            &included_keywords.keywords_pattern,
+                        ))
+                    }
+                    // If has_kws_in_path is Some(true), then always emit the match
+                    if has_kws_in_path.is_some_and(|b| b == true) {
+                        match_emitter.emit(string_match);
+                    }
+                }
+            }
+            return;
+        }
 
         'included_keyword_search: while let Some(included_keyword_match_start) =
             included_keyword_matches.next(regex_caches)